+++ src/be/nikiroo/utils/test/Test.java
@@ -30,7 +30,7 @@ public class Test extends TestLauncher {
 		addSeries(new SerialServerTest(args));
 		addSeries(new StringUtilsTest(args));
 		addSeries(new TempFilesTest(args));
-
+		
 		// TODO: test cache and downloader
 		Cache cache = null;
 		Downloader downloader = null;